Add working scale normalization to training
The scales of the high and low pass parameters after the wavelet decomposition of the
images are different and this means that the loss will give different relative weight
to them. We avoid this by ensuring the scales are similar with the RescalingModule,
which also allows the inverse transformation. The trainer can now pre-transform the
images before passing them to the model. The model trains with the transforms flag set
to false, so that the input and output both use normalized scales. Optionally, the
model transforms flag is set, and then the model can receive as input unnormalized
wavelet decompositions, transform them, perform the encoding/decoding, and then post
transform them so that the output can readily be passed to the inverse wavelet transform.
It is designed this way so that the least number of modifications is used while still
allowing for the option to parametricaly choose whether to scale the channels or not.

* utils.py Contains the new rescaling module class defn
* model.py Has a rescaling module as part of the model now, which can be switched
on or off, in order to pre and post transform the input, in the case when the
training or visualization requires it.
* loggers.py Add a flag which ensures that the model uses the transforms when
required, but does not outside of the logging function.
* train.py Add a flag which controls whether the transforms are used or not.
* trainers/ipe_vae.py Add flag that is controlled from train.py script
* losses.py Add logging of std of channels for debugging and sanity check in
Tensorboard